+++ tests/testthat/test-mutex.r
@@ -2,11 +2,13 @@
 
 test_that("mutex", {
   
-  expect_error(mutex(name = 'abc123', file = tempfile()))
-  
   x <- expect_silent(mutex(cleanup = TRUE))
   y <- expect_no_error(mutex(assert = 'create', file = tempfile()))
   z <- expect_silent(mutex(name = y$name, assert = 'exists'))
+  
+  expect_error(mutex(name = uid(),  file = tempfile()))
+  expect_error(mutex(name = uid(),  assert = 'exists'))
+  expect_error(mutex(name = x$name, assert = 'create'))
   
   expect_true(with(x, invisible(TRUE)))
   
@@ -33,23 +35,47 @@ test_that("mutex", {
   
   mut <- expect_silent(mutex())
   f <- function (nm, shr) {
-    m <- interprocess::mutex(name = nm)
+    m <- interprocess::mutex(name = nm, assert = 'exists')
     return (with(m, TRUE, FALSE, shared = shr, timeout_ms = 0))
   }
   
+  # Use a mutex created by another process.
   expect_true(callr::r(f, list(nm = mut$name, shr = TRUE)))
   expect_true(callr::r(f, list(nm = mut$name, shr = FALSE)))
   
+  # Interact with a locked (shared) mutex.
   with(mut, shared = TRUE, {
     expect_true( callr::r(f, list(nm = mut$name, shr = TRUE)))
     expect_false(callr::r(f, list(nm = mut$name, shr = FALSE)))
   })
   
+  # Interact with a locked (exclusive) mutex.
   with(mut, shared = FALSE, {
     expect_false(callr::r(f, list(nm = mut$name, shr = TRUE)))
     expect_false(callr::r(f, list(nm = mut$name, shr = FALSE)))
   })
   
+  # Mutex unlocked on process exit
+  g <- function (nm, shr) {
+    m <- interprocess::mutex(name = nm, assert = 'exists')
+    return (m$lock(shared = shr, timeout_ms = 0))
+  }
+  expect_true(callr::r(g, list(nm = mut$name, shr = TRUE)))
+  expect_true(with(mut, TRUE, FALSE, timeout_ms = 0))
+  expect_true(callr::r(g, list(nm = mut$name, shr = FALSE)))
+  expect_true(with(mut, TRUE, FALSE, timeout_ms = 0))
+  
+  expect_true(mut$remove())
+  
+  
+  # cleanup works
+  nm  <- callr::r(function () interprocess::mutex(cleanup = TRUE)$name)
+  mut <- expect_silent(interprocess::mutex(name = nm, assert = 'create'))
+  expect_true(mut$remove())
+  
+  # persistence works
+  nm  <- callr::r(function () interprocess::mutex(cleanup = FALSE)$name)
+  mut <- expect_silent(interprocess::mutex(name = nm, assert = 'exists'))
   expect_true(mut$remove())
   
 })