+++ NEWS.md
@@ -1,3 +1,9 @@
+# interprocess 1.2.0
+
+* Renamed `queue()` to `msg_queue()`.
+* Changed length of `uid()` and `hash()` to 11 characters.
+
+
 # interprocess 1.1.0
 
 * Changed length of `uid()` and `hash()` to 12 characters.